requirements.txt
numpy
torch
onnxruntime
stable-baselines3
pygame
gymnasium
